requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.2
pydantic-settings>=2.6.0

# PDF processing
pypdf>=5.1.0
pdfplumber>=0.11.4

# Embeddings & vector store
sentence-transformers>=3.2.1
faiss-cpu>=1.13.2
numpy>=1.26.4

# LLM
anthropic>=0.39.0

# Tokenization for chunking
tiktoken>=0.8.0

# Frontend
streamlit>=1.39.0
requests>=2.32.3

# Utils
python-dotenv>=1.0.1